feat: Add route handling in router.ts
